AMBARI-1108 - PUT call to change the state on host_components collection returns 200 (no op), even though GET with the same predicate returns a number of host_components. (Tom Beerbower via mahadev)
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/trunk@1431843 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-server/src/main/python/bootstrap.py
@@ -106,6 +106,12 @@ class SSH(threading.Thread):
     logFile = open(logFilePath, "a+")
     logFile.write(self.ret["log"])
     logFile.close
+
+    doneFilePath = os.path.join(self.bootdir, self.host + ".done")
+    doneFile = open(doneFilePath, "w+")
+    doneFile.write(str(sshstat.returncode))
+    doneFile.close()
+
     logging.info("Setup agent done for host " + self.host + ", exitcode=" + str(sshstat.returncode))
     pass
 pass
@@ -307,9 +313,10 @@ class BootStrap:
     If .done file for any host is not created, the bootstrap will hang or fail due to timeout"""
     for key, value in self.statuses.iteritems():
       doneFilePath = os.path.join(self.bootdir, key + ".done")
-      doneFile = open(doneFilePath, "w+")
-      doneFile.write(str(value["exitstatus"]))
-      doneFile.close()
+      if not os.path.exists(doneFilePath):
+        doneFile = open(doneFilePath, "w+")
+        doneFile.write(str(value["exitstatus"]))
+        doneFile.close()
     pass
 
   def run(self):
